backend/requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# AWS SDK and Strands Agents
boto3>=1.34.0
strands-agents>=1.0.0
strands-agents-tools>=0.2.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Video Processing
ffmpeg-python>=0.2.0
opencv-python-headless>=4.9.0

# Vector Database
qdrant-client>=1.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.0

# Logging and Monitoring
structlog>=24.1.0
